# Improve performance for normal text