Add benchmark script and improve performance for text without carriage returns:

**master:**
With carriage: 4.329736691100001 ms
Without carriage: 0.24382779939998656 ms

**PR:**
With carriage: 4.286559703500018 ms
Without carriage: 0.003943884600028832 ms